refactor: list, portrait atoms
