+++ src/components/Search/index.tsx
@@ -1,16 +1,8 @@
-import React, {useEffect, useRef, useState} from 'react';
-import {
-	ActivityIndicator,
-	FlatList,
-	SafeAreaView,
-	StyleSheet,
-	TouchableWithoutFeedback,
-	View,
-} from 'react-native';
-import {Button, Image, SearchBar, Text} from 'react-native-elements';
-import RBSheet from 'react-native-raw-bottom-sheet';
+import React, {useEffect, useState} from 'react';
+import {FlatList, SafeAreaView, StyleSheet, View} from 'react-native';
+import {SearchBar, Text} from 'react-native-elements';
 import {getSearchingMovies, getUpcomingMovies} from '../../utils';
-import {SignOutButton} from '../atoms';
+import List from '../atoms/List';
 
 interface Props {
 	navigation: any;
@@ -18,7 +10,8 @@ interface Props {
 	searching: any;
 }
 
-const Header: React.FC<Props> = ({searching}) => {
+const Header: React.FC<Props> = props => {
+	const {searching} = props;
 	const [inputSearch, setInputSearch] = useState('');
 
 	useEffect(() => {
@@ -27,7 +20,7 @@ const Header: React.FC<Props> = ({searching}) => {
 				searching(m.results);
 			})
 			.catch(e => console.log(e));
-	}, [inputSearch]);
+	}, [inputSearch, searching]);
 
 	return (
 		<View style={{backgroundColor: '#000', flex: 1}}>
@@ -48,72 +41,7 @@ const Header: React.FC<Props> = ({searching}) => {
 
 const SearchItem: React.FC<Props> = props => {
 	const {detail, index, navigation} = props;
-	const moviesRef = useRef([]);
-
-	return (
-		<TouchableWithoutFeedback onPress={() => moviesRef.current[index].open()}>
-			<View style={styles.itemList}>
-				<Image
-					source={{
-						uri: `https://www.themoviedb.org/t/p/w220_and_h330_face/${detail.poster_path}`,
-					}}
-					containerStyle={styles.imageList}
-					PlaceholderContent={<ActivityIndicator />}
-				/>
-				<Text style={styles.itemText}>{detail.original_title}</Text>
-				<RBSheet
-					ref={el => (moviesRef.current[index] = el)}
-					closeOnDragDown={true}
-					closeOnPressMask={true}
-					customStyles={{
-						container: {
-							backgroundColor: '#2b2b2b',
-							borderTopLeftRadius: 10,
-							borderTopRightRadius: 10,
-							padding: 10,
-						},
-						wrapper: {
-							backgroundColor: 'rgba(0,0,0,0.2)',
-						},
-						draggableIcon: {
-							backgroundColor: '#000',
-						},
-					}}>
-					<View style={styles.info}>
-						<Image
-							source={{
-								uri: `https://www.themoviedb.org/t/p/w220_and_h330_face/${detail.poster_path}`,
-							}}
-							containerStyle={styles.imageSheet}
-							PlaceholderContent={<ActivityIndicator />}
-						/>
-						<View style={styles.infoDetail}>
-							<Text style={styles.text}>{detail.original_title}</Text>
-							<Text style={styles.text}>{detail.release_date}</Text>
-							<Text style={styles.textOverview}>{detail.overview}</Text>
-							<Text style={styles.text}>{detail.vote_average}</Text>
-							<Button
-								title="Details & More"
-								buttonStyle={styles.buttonStyle}
-								containerStyle={styles.containerButtonStyle}
-								titleStyle={styles.titleButtonStyle}
-								onPress={() =>
-									navigation.navigate('DetailScreen', {
-										id: detail.id,
-										poster_path: detail.poster_path,
-										original_title: detail.original_title,
-										overview: detail.overview,
-										release_date: detail.release_date,
-										vote_average: detail.vote_average,
-									})
-								}
-							/>
-						</View>
-					</View>
-				</RBSheet>
-			</View>
-		</TouchableWithoutFeedback>
-	);
+	return <List navigation={navigation} item={detail} index={index} />;
 };
 
 const Search: React.FC<Props> = ({navigation}) => {
@@ -181,17 +109,6 @@ const styles = StyleSheet.create({
 		fontWeight: '900',
 		fontSize: 20,
 	},
-	imageList: {
-		height: 80,
-		width: 120,
-	},
-	itemList: {
-		backgroundColor: '#21201E',
-		flexDirection: 'row',
-		borderColor: '#000',
-		borderWidth: 1,
-		alignContent: 'center',
-	},
-	itemText: {color: '#cac7c5', textAlignVertical: 'center', margin: 10},
+
 	headerRight: {flexDirection: 'row'},
 });
