+++ src/components/atoms/index.tsx
@@ -1,5 +1,9 @@
 import SearchButton from "./SearchButton";
 import SignOutButton from './SignOutButton'
+import Portrait from './Portrait'
+import List from './List'
 
 export SearchButton from './SearchButton';
 export SignOutButton from './SignOutButton';
+export Portrait from './Portrait';
+export List from './List';
